image-verification-mcp/requirements.txt
fastmcp>=0.2.0
pillow>=10.0.0
pytesseract>=0.3.10
piexif>=1.1.3
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
pdf2image>=1.16.0
pypdf2>=3.0.0
